save record dropbox_files model using paperclip

+++ app/controllers/dropbox_files_controller.rb
@@ -1,9 +1,9 @@
 class DropboxFilesController < ApplicationController
   
   def welcome
+
+      @client = dropbox_client
   
-      @client = Dropbox::API::Client.new(:token => session[:access_token], :secret => session[:access_secret_token])
-      
       @photos=[]  
 
       if @client.ls.any?
@@ -18,23 +18,35 @@ class DropboxFilesController < ApplicationController
 
   def viewmap
 
+      @client = dropbox_client
+
       photo = @client.find('/photos/IMAG0394.jpg')
       destination_file_full_path = Rails.root.to_s + "/" + photo.path.to_s.split("/").last
-      # binding.pry
-      # @image_input = MiniMagick::Image.open(photo.download)
-      # @image_input.write(destination_file_full_path)
-      
+            
       remote_photo = open(photo.direct_url.url,:ssl_verify_mode => OpenSSL::SSL::VERIFY_NONE).read
-      # binding.pry
       # exif_data = EXIFR::JPEG.new(remote_photo)
-      
-      open(destination_file_full_path, 'wb') do |file|
-          file << photo.download
+
+      if File.exists?(destination_file_full_path)
+      else
+        open(destination_file_full_path, 'wb') do |file|
+            file << photo.download
+        end
       end
 
+      if File.exists?(destination_file_full_path)
+         dp_file = File.open(destination_file_full_path)
+      end
+
+      if dp_file
+        DropboxFile.new(:image => dp_file).save!
+      end
+     
       exif_data = EXIFR::JPEG.new(destination_file_full_path)
-      @latitude = exif_data.gps_latitude[0].to_f+exif_data.gps_latitude[1].to_f/60+exif_data.gps_latitude[2].to_f/3600
-      @longitude = -(exif_data.gps_longitude[0].to_f+exif_data.gps_longitude[1].to_f/60+exif_data.gps_longitude[2].to_f/3600)
+
+      binding.pry
+
+      @latitude = exif_data.gps.latitude
+      @longitude = exif_data.gps.longitude
 
       render "viewmap"
     
@@ -63,7 +75,7 @@ class DropboxFilesController < ApplicationController
       if dp_file
         DropboxFile.create(:image => dp_file)
         File.delete(dp_file)
-    end
+      end
   end
   
   def download_file_from_url(obj)
